requirements.txt
pandas
numpy
matplotlib
seaborn
hydra-core
rich
pydantic
tqdm
scikit-learn
statsmodels
typing_extensions